requirements.txt
gradio==5.7.1
hydra-core==1.3.2
lightning==2.4.0
matplotlib==3.8.4
numpy==2.1.3
omegaconf==2.3.0
pandas==1.5.3
PyYAML==5.4.1
PyYAML==6.0.2
regex==2023.12.25
rootutils==1.0.7
scikit_learn==1.4.2
scipy==1.14.1
seaborn==0.13.2
shap==0.45.1
skops==0.9.0
sts_select==0.0.1
tabulate==0.9.0
torch==2.3.1
torcheval==0.0.7
torchmetrics==1.4.0.post0
tqdm==4.66.2
xgboost==2.1.3
